Add files via upload
Air Bnb Data was gotten from Kaggle and this data was cleaned on excel and analysed on Tablaeu app. This data analysis generated useful insight especially for someone thinking of when is the best time to put house up for Airbnb or get an Airbnb, also where would be best favourable, and the exact mapping of the location to the zipcode.
I showed a graph of zip code against the average price of air bnb, also the time graph against the average price, and average price per bedroom. All this inght were created from the this Data. Data were properly cleand in excel then explored